requirements.txt
python-dotenv
openai>=1.0